handle promises witihin useXkcd
Instead of putting this work on the consumer of the hook, handle the async and return only what's relevant to the consumer.

This refactor means that the user of the hook only needs to check whether the state is loading / in error.

If it's not, then it is safe to render.

No need to use a useEffect / useAsync as the hook now does this on behalf of the consumer.