Handle error of 0 in fetchComic
If a hash returns 0, this will result in an error in `fetchComic(0)`.

To address this we _first_ attempt to find a new non-zero result that is unique from all others in the set.

If that fails, we fall back to one.

A _simpler_ alternative would be to abort and just have one fewer comics on days when this happens.

For now, this is a reasonable solution that in very rare cases may result in a duplicative comic, but would not error because it cannot find a result.