Post about turning thirty
This post was a long time in the making. Pretty cool to come back to it months later.

I think it's holding up well. Even better, coming at it fresh after months, I made a much tighter essay.

Certainly more edits would improve it, but it's a nice example of _why_ multiple drafts are useful.