fix: selection should be global
the *+::selection was a typo. It was selecting _adjacent_ to everything.
This resulted in undesireable behavior.

For example, the first paragraph would have a different highlighting structure.

If there was italicized or bold text in the first paragraph, they'd also be different.

This fixes that _without_ being overly specific, allowing code blocks to retain their style.