fix: relative link on list
gatsby uses reach/router under the hood. by default, this means
that it assumes relative links. unfortunately, i no longer want
to use relative links because my notes are all in a flat hierarchy.
the result is that i need to go "up" a level (but this only works
because list is a _sibling_ to my notes - i.e. if i had a note where
the slug was "list" it would be in conflict).