feat: move blog pages to notes
in gatsby-node refocus on using the slug for the note instead of
derving a slug based on the file path.

the slug for all published notes is expected to be in the frontmatter.
if it's not, the build will fail with an error saying what to fix.

this did mean that we needed a new template for notes, so I forked
the current blogEntry to make a notesEntry. they're otherwise
identical.

the blog also now uses the slug,