feat: manual redirects + antilibrary
create a new manual redirect function. the important point here
is that in order to create a manual redirect, the old page needs
to be created... _even if it's never used_.

The `manualRedirect` function handles all of this for and brings
it out of the gatsby-node file to try to minimize clutter.

in this case, i'm using it to create a redirect from a page that
no longer exists - namely the `reading-list` file which was
renamed to anti-library.

the point that is still worth investigating is whether it is
possible ot migrate to using the slug within the frontmatter
more reliably. the hardest part about getting the manual
redirect to work was recognizing that the blogEntry template
still uses the slug generated by the node field -- using the
noteEntry template seems like it would have solved this
problem!

in fact - a quick update to use the notesEntry template does
work as expected. However, it's worth noting that the
assumption then is that _all_ manual redirects will be for
notes entries only - hence the clarification in the function
name.